feat: Concatenate algorithms
